requirements.txt
Flask==3.1.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.2.2
requests==2.32.5
SQLAlchemy==2.0.48
